requirements.txt
# =====================================
# 🤖 OBA Robot - Minimal Requirements
# Hacı Abi'nin Sade ve Etkili Yaklaşımı
# =====================================

# =====================================
# 🔥 CORE DEPENDENCIES (Kritik)
# =====================================
numpy==1.24.3                    # Temel matematik - her yerde kullanılır
pyyaml==6.0.1                    # Config dosyaları için şart

# =====================================
# 🎥 COMPUTER VISION (Minimal)
# =====================================
opencv-python-headless==4.8.1.78 # GUI olmadan OpenCV (600MB tasarruf!)
opencv-contrib-python-headless==4.8.1.78 # AprilTag detection için
Pillow==10.0.0                   # Basit resim işleme

# =====================================
# 🌐 WEB INTERFACE (FastAPI Migration)
# =====================================
fastapi==0.104.1                # Modern async web framework
uvicorn[standard]==0.24.0        # ASGI server
websockets==12.0                 # WebSocket support
jinja2==3.1.2                   # Template engine
python-multipart==0.0.6         # Form data support

# Flask (backward compatibility - remove later)
flask==2.3.2                    # Old web framework
flask-socketio==5.3.5            # Real-time iletişim
python-socketio==5.8.0           # Socket.IO server
eventlet==0.33.3                 # Async networking

# =====================================
# 🔧 HARDWARE (Ortam Bazlı)
# =====================================
# Raspberry Pi donanımı (sadece Linux'ta)
RPi.GPIO==0.7.1; sys_platform == "linux"
gpiozero==1.6.2; sys_platform == "linux"
adafruit-circuitpython-motor==3.4.8; sys_platform == "linux"

# =====================================
# 🚀 ASYNC & COMMUNICATION
# =====================================
asyncio-mqtt==0.13.0             # MQTT client (IoT iletişim)
aiofiles==23.2.1                 # Async file operations

# =====================================
# 📊 UTILITIES (Minimal)
# =====================================
psutil==5.9.5                   # Sistem monitoring
colorlog==6.7.0                 # Renkli logging
pyserial==3.5                   # Serial iletişim

# =====================================
# 🛠️ DEVELOPMENT (Sadece gerekli)
# =====================================
pytest==7.4.0                   # Test framework
black==23.7.0                   # Code formatter

# =====================================
# 📋 FUTURE EXPANSION (Şimdilik kapalı)
# =====================================
# Bu paketler v2.0'da açılabilir:
# scipy==1.11.1                    # Advanced math (50MB)
# matplotlib==3.7.2                # Plotting (100MB)
# pandas==2.0.3                    # Data analysis (80MB)
# scikit-learn==1.3.0              # Machine learning (200MB)
# torch==2.0.1                     # Deep learning (800MB)
# jupyter==1.0.0                   # Notebooks (150MB)
# adafruit-circuitpython-motor==3.4.8  # Motor control
# picamera2==0.3.12                # Pi camera
